# API_documentation.yaml
openapi: 3.0.3
info:
  version: 1.0.0
  title: COURSE API
  description: This is the API Documentation of Course Recommender System of Team Sept-07
servers:
  - url: http://127.0.0.1:5001
...
paths:
  /api/login:
    
    description: End Point for Login
    post:
      tags:
        - User
      summary: "For Login"
      description: Operation to Login
      operationId: login
      requestBody:
        description: Login
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                  example: mdikram88
                password:
                  type: string
                  example: '12345678'
      responses:
        '200':
          description: ''
          content:
            application/json:
              examples:
                Example 1:
                  value: '{"message": "Login successful","user_id": 1}'
        '403':
          description: ''
          content:
            application/json:
              examples:
                Example 1:
                  value: '{"message": "Incorrect password!"}'
        '404':
          description: ''
          content:
            application/json:
              examples:
                Example 1:
                  value: >-
                    {"message": "The Username you entered isn't connected to an
                    account."}
  /api/signup:
    description: End Point for Signup
    post:
      tags:
        - User
      summary: For SignUp
      description: Operation to Sign up
      operationId: signup
      requestBody:
        description: SignUp
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                  example: mdikram88
                email:
                  type: string
                  example: abcd231@gmail.com
                password:
                  type: string
                  example: '12345678'
                name:
                  type: string
                  example: Mohd Ikram
                reset_code:
                  type: string
                  example: aaaa1111
                roll_no:
                  type: string
                  example: 21f282319
                dual_degree:
                  type: boolean
                  example: true
                side_work:
                  type: boolean
                  example: false
                additional_education:
                  type: string
                  example: B.tech
      responses:
        '201':
          description: ''
          content:
            application/json:
              examples:
                Example 1:
                  value: '{"message": "Account created successfully"}'
        '400':
          description: ''
          content:
            application/json:
              examples:
                Example 1:
                  value: '{"message": "Something Went Wrong"}'
        '403':
          description: ''
          content:
            application/json:
              examples:
                Example 1:
                  value: >-
                    {"message": "The username you entered is already connected
                    to an account."}
                Example 2:
                  value: >-
                    {"message": "The email address you entered is already
                    connected to an account."}
                Example 3:
                  value: >-
                    {"message": "The roll_no you entered is already connected to
                    an account."}
  /api/forget_password:
    description: End Point for Forget Password
    post:
      tags:
        - User
      summary: For Forget Password
      description: Operation to Forget Password
      operationId: forget_password
      requestBody:
        description: Forget Password
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                  example: mdikram88
                reset_code:
                  type: string
                  example: aaaa1111
                password:
                  type: string
                  example: '12345678'
      responses:
        '202':
          description: Accepted
          content:
            application/json:
              examples:
                Example 1:
                  value: '{"message": "Password Reset Successfully"}'
        '403':
          description: ''
          content:
            application/json:
              examples:
                Example 1:
                  value: '{"message": "The reset code you entered is incorrect."}'
        '404':
          description: ''
          content:
            application/json:
              examples:
                Example 1:
                  value: '{"message": "Account doesn''t exist"}'
  /api/student/approve:
    description: End Point for Approve User
    put:
      tags:
        - Admin
      summary: For approving new students
      description: Operation to Approve User
      operationId: approve_user
      requestBody:
        description: Approve User
        content:
          application/json:
            schema:
              type: object
              properties:
                user_id:
                  type: integer
                  example: 1
      responses:
        '200':
          description: ''
          content:
            application/json:
              examples:
                Example 1:
                  value: '{"message": "User approved"}'
        '400':
          description: ''
          content:
            application/json:
              examples:
                Example 1:
                  value: '{"message": "Something went wrong"}'
        '404':
          description: ''
          content:
            application/json:
              examples:
                Example 1:
                  value: '{"message": "invalid user_id"}'
  /api/profile:
    description: End Point for User
    post:
      tags:
        - User
      summary: For creating new User
      description: Operation to Create new User Profile
      operationId: CreateUser
      requestBody:
        description: SignUp
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                  example: mdikram88
                email:
                  type: string
                  example: abcd231@gmail.com
                password:
                  type: string
                  example: '12345678'
                name:
                  type: string
                  example: Mohd Ikram
                reset_code:
                  type: string
                  example: aaaa1111
                roll_no:
                  type: string
                  example: 21f282319
                dual_degree:
                  type: boolean
                  example: true
                side_work:
                  type: boolean
                  example: false
                additional_education:
                  type: string
                  example: B.tech
      responses:
        '201':
          description: ''
          content:
            application/json:
              examples:
                Example 1:
                  value: '{"message": "Account created successfully"}'
        '400':
          description: ''
          content:
            application/json:
              examples:
                Example 1:
                  value: '{"message": "Something Went Wrong"}'
        '403':
          description: ''
          content:
            application/json:
              examples:
                Example 1:
                  value: >-
                    {"message": "The username you entered is already connected
                    to an account."}
                Example 2:
                  value: >-
                    {"message": "The email address you entered is already
                    connected to an account."}
                Example 3:
                  value: >-
                    {"message": "The roll_no you entered is already connected to
                    an account."}
  /api/profile/{user_id}:
    description: End point to read User and modify.
    get:
      tags:
        - User
      summary: For getting User data
      description: Operation to get User details.
      parameters:
        - in: path
          name: user_id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: ''
          content:
            application/json:
              examples:
                Example 1:
                  value: >-
                    { "data": { "dual_degree": false, "email": "abcd@gmail.com",
                    "name": "M.Ikram", "password": "12345678", "reset_code":
                    "12345678", "side_work": false, "user_id": 1, "username":
                    "mdikram88", "verified": true } }
        '404':
          description: ''
          content:
            application/json:
              examples:
                Example 1:
                  value: '{"message": "Invalid user_id"}'
    put:
      tags:
        - User
      summary: For Updating User data
      description: Operation to Update User details.
      parameters:
        - in: path
          name: user_id
          required: true
          schema:
            type: integer
      requestBody:
        description: Update User profile
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                  example: mdikram88
                email:
                  type: string
                  example: abcd231@gmail.com
                password:
                  type: string
                  example: '12345678'
                name:
                  type: string
                  example: Mohd Ikram
                reset_code:
                  type: string
                  example: aaaa1111
                dual_degree:
                  type: boolean
                  example: true
                side_work:
                  type: boolean
                  example: false
                additional_education:
                  type: string
                  example: B.tech
      responses:
        '202':
          description: ''
          content:
            application/json:
              examples:
                Example 1:
                  value: '{"message": "Profile updated successfully"}'
        '400':
          description: ''
          content:
            application/json:
              examples:
                Example 1:
                  value: '{"message": "Invalid User Id"}'
                Example 2:
                  value: '{"message": "Email is already taken, please try another"}'
        '500':
          description: ''
          content:
            application/json:
              examples:
                Example 1:
                  value: '{"message": "Something Went Wrong"}'
...